Move plugins to own package
